requirements.txt
python-dotenv
pinecone-client
langchain
pypdf